Add a btserial command and move serial device rules
In RaspiOS, the udev rules for configuring the /dev/serialN devices
exist in /etc/udev/rules.d/99-com.rules and involve a few chunks of
shell-script for determining which UART to point the serialN alias at.

With the version of udev in Bullseye (and coincidentally, Ubuntu Hirsute
onwards), udev complains about the $ALIASES substitution in this rule. I
see in Bullseye this was fixed by escaping it to $$ALIASES, but I
thought it might be a little cleaner to spin this logic out into its own
btserial script, and stuff all the rules and logic in the pi-bluetooth
package.